extract stardust summary
