+++ packages/desktop/views/dashboard/send-flow/views/components/index.js
@@ -1,4 +1,6 @@
 export { default as DateTimePicker } from './DateTimePicker.svelte'
 export { default as DateTimePickerButton } from './DateTimePickerButton.svelte'
 export { default as DateTimePickerMenu } from './DateTimePickerMenu.svelte'
+export { default as EvmTransactionSummary } from './EvmTransactionSummary.svelte'
+export { default as StardustTransactionSummary } from './StardustTransactionSummary.svelte'
 export { default as TransactionDetails } from './TransactionDetails.svelte'
