+++ packages/shared/src/lib/core/wallet/utils/index.ts
@@ -24,6 +24,7 @@ export * from './isAccountAddress'
 export * from './isReservedTagKeyword'
 export * from './isSubjectInternal'
 export * from './parseGovernanceMetadata'
+export * from './prepareOutputFromTransactionData'
 export * from './processAndAddToActivities'
 export * from './validateIrc30Metadata'
 export * from './validateTokenAmount'
